fix: correct header guard and type mismatch issues

+++ include/GateServos.h
@@ -72,7 +72,7 @@
       void testServo(int servopin);         // Test given servo pin (debug function)
       const int num_gates = NUM_GATES;      //
       int curopengate = -1;                 // cuurrently open gate selected manually with button
-      const int opendelay = OPEN_DELAY;     // ms delay to allow servo to completely open gate
+      const unsigned long opendelay = OPEN_DELAY;     // ms delay to allow servo to completely open gate
       bool gateopen[8] = {false, false, false, false,false, false, false, false};   // array indicating which gates are open
       bool isGateDisabled(int gatenum);     // Check if a gate is disabled (servo pin = -1)
   };
